+++ src/constants.ts
@@ -1,0 +1,9 @@
+export const COMMANDS = {
+    OPEN_RELATED_FILE: 'els.openRelatedFile',
+    RELOAD_PROJECT: 'els.reloadProject',
+    SET_STATUS_BAR_TEXT: 'els.setStatusBarText',
+    RUN_IN_EMBER_CLI: 'els.runInEmberCLI',
+    EXECUTE_IN_EMBER_CLI: 'els.executeInEmberCLI',
+    SET_CONFIG: 'els.setConfig',
+    GET_RELATED_FILES: 'els.getRelatedFiles'
+}
